+++ spec/shapes/hash_spec.rb
@@ -48,8 +48,54 @@ RSpec.describe Shaped::Shapes::Hash do
         end
       end
 
-      context 'when the test hash does not match the specified shape' do
+      context 'when the test hash has an Integer email value' do
         let(:test_hash) { { email: 2_048 } }
+
+        it 'returns false' do
+          expect(matched_by?).to eq(false)
+        end
+      end
+
+      context 'when the test hash has an extra key (`:name`)' do
+        let(:test_hash) { { email: 'jim@gmail.com', name: 'Jimbo' } }
+
+        it 'returns true' do
+          expect(matched_by?).to eq(true)
+        end
+      end
+    end
+
+    context 'when the shape description is `{ email: Shaped::Shape(String, NilClass) }`' do
+      subject(:hash_shape_description) { { email: Shaped::Shape(String, nil) } }
+
+      context "when the test hash is { email: 'jim@gmail.com' }" do
+        let(:test_hash) { { email: 'jim@gmail.com' } }
+
+        it 'returns true' do
+          expect(matched_by?).to eq(true)
+        end
+      end
+
+      context 'when the test hash is { email: nil }' do
+        let(:test_hash) { { email: nil } }
+
+        it 'returns true' do
+          expect(matched_by?).to eq(true)
+        end
+      end
+
+      context 'when the test hash is an empty hash' do
+        # although the :email key is not given explicitly, the
+        # value is effectively `nil`, which is an allowed value.
+        let(:test_hash) { {} }
+
+        it 'returns true' do
+          expect(matched_by?).to eq(true)
+        end
+      end
+
+      context 'when the test hash is { email: 951 }' do
+        let(:test_hash) { { email: 951 } }
 
         it 'returns false' do
           expect(matched_by?).to eq(false)
@@ -63,6 +109,14 @@ RSpec.describe Shaped::Shapes::Hash do
 
     it 'returns a readably formatted description of the expected hash shape' do
       expect(to_s).to eq('{ :email => String }')
+    end
+
+    context 'when an expected value is a specific string' do
+      let(:hash_shape_description) { { secret_key: 'ABC123' } }
+
+      it 'renders the string within quotes' do
+        expect(to_s).to eq('{ :secret_key => "ABC123" }')
+      end
     end
   end
 end